🚚 move pages and styles in sources folder
