+++ sources/components/Board.jsx
@@ -9,19 +9,20 @@ import parse from "html-react-parser";
 
 import { SlideContext } from "../context/SlideContext";
 
-import "../../styles/slides.css";
+import "../styles/slides.css";
 import "../../node_modules/reveal.js/dist/reveal.css";
 import "../../node_modules/reveal.js/dist/theme/night.css";
 
 import { Stack } from "@mui/material";
-import Button from '@mui/material/Button';
+import Button from "@mui/material/Button";
 
 import { uid } from "uid";
 
 const Board = () => {
   const { roomId } = useParams();
 
-  const { slideNumber, setSlideNumber, slideTotal, setSlideTotal } = useContext(SlideContext);
+  const { slideNumber, setSlideNumber, slideTotal, setSlideTotal } =
+    useContext(SlideContext);
 
   const [slides, setSlides] = useState([]);
   const [revealInitialize, setRevealInitialize] = useState(false);
@@ -69,14 +70,14 @@ const Board = () => {
     Reveal.destroy();
     setRevealInitialize(false);
   }, [slideTotal, slides]);
-  
+
   const deletePage = useCallback(() => {
-    set(ref(db, `room/${roomId}/slide/${slideTotal-1}`), null);
+    set(ref(db, `room/${roomId}/slide/${slideTotal - 1}`), null);
     setSlideTotal(slideTotal - 1);
     Reveal.destroy();
     setRevealInitialize(false);
-  } , [slideNumber, slideTotal]);
-  
+  }, [slideNumber, slideTotal]);
+
   return (
     <>
       <Stack direction="row" className="add-buttons" spacing={2}>
@@ -84,9 +85,9 @@ const Board = () => {
           Ajouter une page
         </Button>
         {slideNumber > 0 && (
-        <Button variant="contained" color="error" onClick={deletePage}>
-          Supprimer la dernière page
-        </Button>
+          <Button variant="contained" color="error" onClick={deletePage}>
+            Supprimer la dernière page
+          </Button>
         )}
       </Stack>
       <div className="reveal">
