+++ sources/pages/Login.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { auth } from "../sources/services/firebase-config";
+import { auth } from "../services/firebase-config";
 import {
   signInWithPopup,
   GoogleAuthProvider,
@@ -29,7 +29,7 @@ const Login = () => {
         break;
     }
     signInWithPopup(auth, provider);
-    console.log(auth)
+    console.log(auth);
   };
 
   return (
