+++ sources/pages/Preview.jsx
@@ -1,26 +1,27 @@
 import React, { useEffect, useContext, useState, useCallback } from "react";
 import { useParams } from "react-router-dom";
 
-import { db } from "../sources/services/firebase-config";
+import { db } from "../services/firebase-config";
 import { ref, set, onValue, child, update, push } from "firebase/database";
 
 import Reveal from "reveal.js";
 import parse from "html-react-parser";
 
-import { SlideContext } from "../sources/context/SlideContext";
+import { SlideContext } from "../context/SlideContext";
 
-import "../node_modules/reveal.js/dist/reveal.css";
-import "../node_modules/reveal.js/dist/theme/night.css";
+import "../../node_modules/reveal.js/dist/reveal.css";
+import "../../node_modules/reveal.js/dist/theme/night.css";
 
 import { Stack } from "@mui/material";
-import Button from '@mui/material/Button';
+import Button from "@mui/material/Button";
 
 import { uid } from "uid";
 
 const Preview = () => {
   const { roomId } = useParams();
 
-  const { slideNumber, setSlideNumber, slideTotal, setSlideTotal } = useContext(SlideContext);
+  const { slideNumber, setSlideNumber, slideTotal, setSlideTotal } =
+    useContext(SlideContext);
 
   const [slides, setSlides] = useState([]);
   const [revealInitialize, setRevealInitialize] = useState(false);
@@ -65,17 +66,17 @@ const Preview = () => {
     Reveal.destroy();
     setRevealInitialize(false);
   }, [slideTotal, slides]);
-  
+
   const deletePage = useCallback(() => {
-    set(ref(db, `room/${roomId}/slide/${slideTotal-1}`), null);
+    set(ref(db, `room/${roomId}/slide/${slideTotal - 1}`), null);
     setSlideTotal(slideTotal - 1);
     Reveal.destroy();
     setRevealInitialize(false);
-  } , [slideNumber, slideTotal]);
-  
+  }, [slideNumber, slideTotal]);
+
   return (
     <>
-      <div className="reveal" style={{width: "100vw", height: "100vh"}}>
+      <div className="reveal" style={{ width: "100vw", height: "100vh" }}>
         <div className="slides">
           {slides.length !== 0 &&
             slides.map((slide, index) => (
